I've Failed Chapter03 of Kevin's book.
Pretty infuriating right now. You'd think I'd be able to do a simple
hello world program in openGL by now... It's been over a week I think...

And yet. Rendering triangles is still NOT a trivial task for me.

That 20hour rule that ukellali playing fool mentioned on that ted
talk... Yeah. I don't think it applies to programming. Either that or I
am a moron.